tests/end2end/helpers: add Confirm
