+++ tests/end2end/document/document/UC09_delete_section/UC09_T01_delete_section/test_UC09_T01_delete_section.py
@@ -1,6 +1,7 @@
 from seleniumbase import BaseCase
 
 from tests.end2end.end2end_test_setup import End2EndTestSetup
+from tests.end2end.helpers.components.confirm import Confirm
 from tests.end2end.helpers.screens.document_tree.screen_document_tree import (
     Screen_DocumentTree,
 )
@@ -30,9 +31,10 @@ class Test_UC09_T01_DeleteSection(BaseCase):
 
             screen_document.assert_node_title_contains("First section")
 
-            screen_document.do_node_delete()
-            screen_document.assert_confirm_section_delete()
-            screen_document.do_confirm_action()
+            confirm: Confirm = screen_document.do_node_delete()
+            confirm.assert_confirm()
+            confirm.assert_confirm_section_delete()
+            confirm.do_confirm_action()
 
             screen_document.assert_no_text("First section")
             screen_document.assert_toc_contains_not("First section")
