+++ tests/end2end/document/document/UC10_delete_requirement/UC10_T01_delete_requirement/test_UC10_T01_delete_requirement.py
@@ -1,6 +1,7 @@
 from seleniumbase import BaseCase
 
 from tests.end2end.end2end_test_setup import End2EndTestSetup
+from tests.end2end.helpers.components.confirm import Confirm
 from tests.end2end.helpers.screens.document_tree.screen_document_tree import (
     Screen_DocumentTree,
 )
@@ -30,9 +31,10 @@ class Test_UC10_T01_DeleteRequirement(BaseCase):
 
             screen_document.assert_node_title_contains("Requirement title")
 
-            screen_document.do_node_delete()
-            screen_document.assert_confirm_requirement_delete()
-            screen_document.do_confirm_action()
+            confirm: Confirm = screen_document.do_node_delete()
+            confirm.assert_confirm()
+            confirm.assert_confirm_requirement_delete()
+            confirm.do_confirm_action()
 
             screen_document.assert_no_text("Requirement title")
             screen_document.assert_toc_contains_not("Requirement title")
